+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-# Fixed issues
+# Changelog
+ - code/win32_handmade.cpp  82 : Create function to close window.
  - code/win32_handmade.cpp  34 : This won't be needed anymore - addresses #3